# Vectorize `rotate` even better