@StephanTLavavej 2025-05-22
When things spin faster, science happens faster! :rocket: :stopwatch: :joy_cat: